CorrectionBasedBear: Move to bearlib.abstractions
