+++ bears/c_languages/IndentBear.py
@@ -1,5 +1,6 @@
 import platform
-from bears.linters.CorrectionBasedBear import CorrectionBasedBear
+
+from coalib.bearlib.abstractions.CorrectionBasedBear import CorrectionBasedBear
 
 
 class IndentBear(CorrectionBasedBear):
