--- bears/linters/__init__.py
@@ -1,3 +1,0 @@
-"""
-This package contains wrapper bears around usual linters.
-"""
