+++ coalib/bearlib/abstractions/CorrectionBasedBear.py
