+++ coalib/tests/bearlib/abstractions/CorrectionBasedBearTest.py
